refactor: 지도 스크립트 추가와 장소 검색 로직 분리
