+++ src/hooks/useMapPlaces.ts
@@ -1,22 +1,19 @@
 import { useEffect, useState } from 'react';
 
+import useMapScript from './useMapScript';
+
 const useMapPlaces = () => {
   const [places, setPlaces] = useState<kakao.maps.services.Places>();
   const [result, setResult] = useState<kakao.maps.services.PlacesSearchResult>();
+  const { script } = useMapScript();
 
   useEffect(() => {
-    const script = document.createElement('script');
-    script.type = 'text/javascript';
-    script.src = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${
-      import.meta.env.VITE_KAKAO_API_JS_KEY
-    }&libraries=services&autoload=false`;
-
-    document.head.appendChild(script);
+    if (!script) return;
 
     script.addEventListener('load', () => {
       kakao.maps.load(() => setPlaces(new kakao.maps.services.Places()));
     });
-  }, []);
+  }, [script]);
 
   const searchPlaces = (keyword: string) => {
     if (!places) return;
